Extract converters to fit ckeditor 5 pattern
